requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
distro==1.6.0
Django==3.2.9
djangorestframework==3.12.4
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonschema==3.2.0
kombu==5.2.2
msgpack==1.0.3
oauthlib==3.1.1
paramiko==2.8.0
prompt-toolkit==3.0.22
psycopg2==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
redis==4.0.2
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
websockets==10.1
wrapt==1.13.3
zope.interface==5.4.0
